+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix compatibility with Laravel <= 6 of the `sentry:test` and `sentry:publish` commands
+
 ## 2.10.0
 
 - Improve output and DX for `sentry:test` and `sentry:publish` commands (#522)